Fix remaining clang-tidy stragglers from the T13 audit
Addresses the 19 sites left unfixed by the two earlier T13 batches
(tracked in tmp/codex_t13_verify.log): a redundant `protected:`
label, two non-exhaustive-looking switches over `% N` that clang-tidy
can't prove exhaustive, an uninitialized `std::tm`, a dead local that
should have been asserted on (which uncovered that `RefObjectCleanupHonorsAllThreeConditions`
never exercised its own "newest log is also deletable" case), two
confusable-identifier renames, an unused using-declaration, three
named-exception-throw sites annotated with the established
NOLINT(cert-err09-cpp,cert-err61-cpp,misc-throw-by-value-catch-by-reference)
pattern already used elsewhere in the tree, two member functions
made `const`, a `&vector[0]` replaced with `.data()`, two redundant
`static_cast<uint64_t>`, and a missing `const auto *`.

The two `bugprone-use-after-move` findings in gtest_cas_event_log.cpp
got different treatment after inspection: `PoolEmitsToSink` reused an
already-moved-from event for a second, unrelated `emitEvent` call, so
it now constructs a fresh event instead; `EmitEventMovesSourceIntoSink`
reads the moved-from state as its entire point (proving `emitEvent`
truly moves rather than copies), verified empirically to hold on this
toolchain, so it keeps the read and gets a NOLINT with a reason
instead.

Verified: ninja -C build unit_tests_dbms is clean, and the CAS gtest
battery (Cas*:CA*:ContentAddressedLog*:CountingBackendShape*:RefSnapshotCodec*:RefTableCacheEviction*:RefWriter*)
passes 920/920.

+++ src/Disks/DiskObjectStorage/MetadataStorages/ContentAddressed/Backend/CasObjectStorageBackend.cpp
@@ -437,7 +437,7 @@ static String readObjectRanged(IObjectStorage & object_storage, const String & p
     /// Native callers already HEAD the key, so passing its size avoids another metadata round trip.
     /// A zero size means the caller does not know it and metadata must be fetched here.
     const uint64_t object_size = known_size != 0 ? known_size
-        : static_cast<uint64_t>(object_storage.getObjectMetadata(path, /*with_tags=*/false).size_bytes);
+        : object_storage.getObjectMetadata(path, /*with_tags=*/false).size_bytes;
     if (range.offset >= object_size)
         return {};
 
@@ -476,7 +476,7 @@ static std::unique_ptr<ReadBuffer> openObjectRangedStream(IObjectStorage & objec
     /// As in `readObjectRanged`, a caller-supplied size avoids another metadata round trip; zero means
     /// that the size is unknown and must be fetched.
     const uint64_t object_size = known_size != 0 ? known_size
-        : static_cast<uint64_t>(object_storage.getObjectMetadata(path, /*with_tags=*/false).size_bytes);
+        : object_storage.getObjectMetadata(path, /*with_tags=*/false).size_bytes;
     if (range.offset >= object_size)
         return std::make_unique<ReadBufferFromString>(std::string_view{});
